@@ -1,8 +1,5 @@
 module github.com/kinbiko/bugsnag
 
-go 1.12
+go 1.13
 
-require (
-	github.com/kinbiko/jsonassert v1.0.1
-	github.com/pkg/errors v0.8.1
-)
+require github.com/kinbiko/jsonassert v1.0.1
